+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### New features and optimizations
 - Added public-facing official Dockerfile for NeMo-Aligner
 - Memory optimization in PPO that helps avoid OOM in the actor when sending training data to the critic
+- SFT: added support for custom validation metrics based on model generations
 
 ### Breaking changes
 
@@ -16,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 a dictionary from the training configuration.
 - `exp_manager.max_time_per_run` is now respected, the trainers will save and run validation before exiting if we've reached the time limit.
 - Fixed crash in PPO when using a separate reward model server (i.e., with `combine_rm_and_critic_server=False`).
+- Fixed a crash when LR scheduler was not specified
 
 ## [0.1.0] - 2023-12-04
 ### Added